Rename files to be consistent
